Harmonize packages in `mail` module
* Move inbound channel adapters into a new `inbound` package
* Move outbound channel adapters into a new `outbound` package
* Preserve existing classes in the root package as deprecated
and as extensions of new classes
