+++ spring-integration-mail/src/main/java/org/springframework/integration/mail/ImapMailReceiver.java
@@ -16,31 +16,11 @@
 
 package org.springframework.integration.mail;
 
-import java.time.Instant;
-import java.util.Arrays;
-import java.util.Locale;
-import java.util.Objects;
-import java.util.Properties;
-import java.util.concurrent.ScheduledFuture;
-
-import jakarta.mail.Flags;
-import jakarta.mail.Flags.Flag;
-import jakarta.mail.Folder;
-import jakarta.mail.Message;
-import jakarta.mail.MessagingException;
-import jakarta.mail.search.AndTerm;
-import jakarta.mail.search.FlagTerm;
-import jakarta.mail.search.NotTerm;
-import jakarta.mail.search.SearchTerm;
 import org.eclipse.angus.mail.imap.IMAPFolder;
 import org.jspecify.annotations.Nullable;
 
-import org.springframework.scheduling.TaskScheduler;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
-import org.springframework.util.Assert;
-
 /**
- * A {@link MailReceiver} implementation for receiving mail messages from a
+ * A {@link org.springframework.integration.mail.inbound.MailReceiver} implementation for receiving mail messages from a
  * mail server that supports the IMAP protocol. In addition to the pollable
  * {@link #receive()} method, the {@link #waitForNewMessages()} method provides
  * the option of blocking until new messages are available prior to calling
@@ -53,302 +33,17 @@ import org.springframework.util.Assert;
  * @author Gary Russell
  * @author Artem Bilan
  * @author Alexander Pinske
+ *
+ * @deprecated since 7.0 in favor of {@link org.springframework.integration.mail.inbound.ImapMailReceiver}
  */
-public class ImapMailReceiver extends AbstractMailReceiver {
+@Deprecated(forRemoval = true, since = "7.0")
+public class ImapMailReceiver extends org.springframework.integration.mail.inbound.ImapMailReceiver {
 
-	private static final int DEFAULT_CANCEL_IDLE_INTERVAL = 120000;
-
-	private static final String PROTOCOL = "imap";
-
-	private final IdleCanceler idleCanceler = new IdleCanceler();
-
-	private boolean shouldMarkMessagesAsRead = true;
-
-	private SearchTermStrategy searchTermStrategy = new DefaultSearchTermStrategy();
-
-	private long cancelIdleInterval = DEFAULT_CANCEL_IDLE_INTERVAL;
-
-	@SuppressWarnings("NullAway.Init")
-	private TaskScheduler scheduler;
-
-	private boolean isInternalScheduler;
-
-	private volatile @Nullable ScheduledFuture<?> pingTask;
-
-	@SuppressWarnings("this-escape")
 	public ImapMailReceiver() {
-		setProtocol(PROTOCOL);
 	}
 
-	@SuppressWarnings("this-escape")
 	public ImapMailReceiver(@Nullable String url) {
 		super(url);
-		if (url != null) {
-			Assert.isTrue(url.toLowerCase(Locale.ROOT).startsWith(PROTOCOL),
-					"URL must start with 'imap' for the IMAP Mail receiver.");
-		}
-		else {
-			setProtocol(PROTOCOL);
-		}
-	}
-
-	/**
-	 * Check if messages should be marked as read.
-	 * @return true if messages should be marked as read.
-	 */
-	public Boolean isShouldMarkMessagesAsRead() {
-		return this.shouldMarkMessagesAsRead;
-	}
-
-	/**
-	 * Provides a way to set custom {@link SearchTermStrategy} to compile a {@link SearchTerm}
-	 * to be applied when retrieving mail.
-	 * @param searchTermStrategy The search term strategy implementation.
-	 */
-	public void setSearchTermStrategy(SearchTermStrategy searchTermStrategy) {
-		Assert.notNull(searchTermStrategy, "'searchTermStrategy' must not be null");
-		this.searchTermStrategy = searchTermStrategy;
-	}
-
-	/**
-	 * Specify if messages should be marked as read.
-	 * @param shouldMarkMessagesAsRead true if messages should be marked as read.
-	 */
-	public void setShouldMarkMessagesAsRead(Boolean shouldMarkMessagesAsRead) {
-		this.shouldMarkMessagesAsRead = shouldMarkMessagesAsRead;
-	}
-
-	/**
-	 * IDLE commands will be terminated after this interval; useful in cases where a connection
-	 * might be silently dropped. A new IDLE will usually immediately be processed. Specified
-	 * in seconds; default 120 (2 minutes). RFC 2177 recommends an interval no larger than 29 minutes.
-	 * @param cancelIdleInterval the cancelIdleInterval to set
-	 * @since 3.0.5
-	 */
-	public void setCancelIdleInterval(long cancelIdleInterval) {
-		this.cancelIdleInterval = cancelIdleInterval * 1000; // NOSONAR - convert seconds to milliseconds
-	}
-
-	@Override
-	protected void onInit() {
-		super.onInit();
-		this.scheduler = getTaskScheduler();
-		if (this.scheduler == null) {
-			ThreadPoolTaskScheduler tpts = new ThreadPoolTaskScheduler();
-			tpts.initialize();
-			this.scheduler = tpts;
-			this.isInternalScheduler = true;
-		}
-		Properties javaMailProperties = getJavaMailProperties();
-		for (String name : new String[] {PROTOCOL, "imaps"}) {
-			String peek = "mail." + name + ".peek";
-			if (javaMailProperties.getProperty(peek) == null) {
-				javaMailProperties.setProperty(peek, "true");
-			}
-		}
-	}
-
-	@Override
-	public void destroy() {
-		super.destroy();
-		if (this.isInternalScheduler) {
-			((ThreadPoolTaskScheduler) this.scheduler).shutdown();
-		}
-		cancelPing();
-	}
-
-	/**
-	 * The hook to be called when we need to cancel the current ping task and close the mail folder.
-	 * In other words: when IMAP idle should be stopped for some reason.
-	 * The next {@link #waitForNewMessages()} call will re-open the folder and start a new ping task.
-	 * @since 5.2
-	 */
-	public void cancelPing() {
-		if (this.pingTask != null) {
-			this.pingTask.cancel(true);
-		}
-		closeFolder();
-	}
-
-	/**
-	 * This method is unique to the IMAP receiver and only works if IMAP IDLE
-	 * is supported (see RFC 2177 for more detail).
-	 * @throws MessagingException Any MessagingException.
-	 */
-	public void waitForNewMessages() throws MessagingException {
-		openFolder();
-		Folder folder = getFolder();
-		Assert.state(folder instanceof IMAPFolder,
-				() -> "folder is not an instance of [" + IMAPFolder.class.getName() + "]");
-		IMAPFolder imapFolder = (IMAPFolder) folder;
-		if (imapFolder.hasNewMessages()) {
-			return;
-		}
-		else if (!folder.getPermanentFlags().contains(Flags.Flag.RECENT) && searchForNewMessages().length > 0) {
-			return;
-		}
-		try {
-			this.pingTask =
-					this.scheduler.schedule(this.idleCanceler, Instant.now().plusMillis(this.cancelIdleInterval));
-			if (imapFolder.isOpen()) {
-				imapFolder.idle(true);
-			}
-		}
-		finally {
-			if (this.pingTask != null) {
-				this.pingTask.cancel(true);
-			}
-		}
-	}
-
-	/**
-	 * Retrieves new messages from this receiver's folder. This implementation
-	 * creates a {@link SearchTerm} that searches for all messages in the
-	 * folder that are {@link jakarta.mail.Flags.Flag#RECENT RECENT}, not
-	 * {@link jakarta.mail.Flags.Flag#ANSWERED ANSWERED}, and not
-	 * {@link jakarta.mail.Flags.Flag#DELETED DELETED}. The search term is used
-	 * to {@link Folder#search(SearchTerm) search} for new messages.
-	 * @return the new messages
-	 * @throws MessagingException in case of JavaMail errors
-	 */
-	@Override
-	protected Message[] searchForNewMessages() throws MessagingException {
-		Folder folderToUse = getFolder();
-		Assert.state(folderToUse != null, "'folderToUse' should not be null");
-		Flags supportedFlags = folderToUse.getPermanentFlags();
-		SearchTerm searchTerm = compileSearchTerms(supportedFlags);
-		if (folderToUse.isOpen()) {
-			return nullSafeMessages(searchTerm != null ? folderToUse.search(searchTerm) : folderToUse.getMessages());
-		}
-		throw new MessagingException("Folder is closed");
-	}
-
-	private Message[] nullSafeMessages(Message[] messageArray) {
-		boolean hasNulls = false;
-		for (Message message : messageArray) {
-			if (message == null) {
-				hasNulls = true;
-				break;
-			}
-		}
-		if (!hasNulls) {
-			return messageArray;
-		}
-		else {
-			return Arrays.stream(messageArray)
-					.filter(Objects::nonNull)
-					.toArray(Message[]::new);
-		}
-	}
-
-	private @Nullable SearchTerm compileSearchTerms(Flags supportedFlags) {
-		Folder folderToUse = this.getFolder();
-		Assert.state(folderToUse != null, "'folderToUse' should not be null");
-		return this.searchTermStrategy.generateSearchTerm(supportedFlags, folderToUse);
-	}
-
-	@Override
-	protected void setAdditionalFlags(Message message) throws MessagingException {
-		super.setAdditionalFlags(message);
-		if (this.shouldMarkMessagesAsRead) {
-			message.setFlag(Flag.SEEN, true);
-		}
-	}
-
-	private class IdleCanceler implements Runnable {
-
-		IdleCanceler() {
-		}
-
-		@Override
-		public void run() {
-			try {
-				Folder folder = getFolder();
-				logger.debug("Canceling IDLE");
-				if (folder != null) {
-					folder.isOpen(); // resets idle state
-				}
-			}
-			catch (Exception ex) {
-				logger.error(ex, "Error during resetting idle state.");
-			}
-		}
-
-	}
-
-	private class DefaultSearchTermStrategy implements SearchTermStrategy {
-
-		DefaultSearchTermStrategy() {
-		}
-
-		@Override
-		public @Nullable SearchTerm generateSearchTerm(Flags supportedFlags, Folder folder) {
-			SearchTerm searchTerm = null;
-			boolean recentFlagSupported = false;
-			if (supportedFlags != null) {
-				recentFlagSupported = supportedFlags.contains(Flags.Flag.RECENT);
-				if (recentFlagSupported) {
-					searchTerm = new FlagTerm(new Flags(Flags.Flag.RECENT), true);
-				}
-				if (supportedFlags.contains(Flags.Flag.ANSWERED)) {
-					NotTerm notAnswered = new NotTerm(new FlagTerm(new Flags(Flags.Flag.ANSWERED), true));
-					if (searchTerm == null) {
-						searchTerm = notAnswered;
-					}
-					else {
-						searchTerm = new AndTerm(searchTerm, notAnswered);
-					}
-				}
-				if (supportedFlags.contains(Flags.Flag.DELETED)) {
-					NotTerm notDeleted = new NotTerm(new FlagTerm(new Flags(Flags.Flag.DELETED), true));
-					if (searchTerm == null) {
-						searchTerm = notDeleted;
-					}
-					else {
-						searchTerm = new AndTerm(searchTerm, notDeleted);
-					}
-				}
-				if (supportedFlags.contains(Flags.Flag.SEEN)) {
-					NotTerm notSeen = new NotTerm(new FlagTerm(new Flags(Flags.Flag.SEEN), true));
-					if (searchTerm == null) {
-						searchTerm = notSeen;
-					}
-					else {
-						searchTerm = new AndTerm(searchTerm, notSeen);
-					}
-				}
-			}
-
-			if (!recentFlagSupported) {
-				searchTerm = applyTermsWhenNoRecentFlag(folder, searchTerm);
-			}
-			return searchTerm;
-		}
-
-		private SearchTerm applyTermsWhenNoRecentFlag(Folder folder, @Nullable SearchTerm searchTerm) {
-			NotTerm notFlagged;
-			if (folder.getPermanentFlags().contains(Flag.USER)) {
-				logger.debug(() -> "This email server does not support RECENT flag, but it does support " +
-						"USER flags which will be used to prevent duplicates during email fetch." +
-						" This receiver instance uses flag: " + getUserFlag());
-				Flags siFlags = new Flags();
-				siFlags.add(getUserFlag());
-				notFlagged = new NotTerm(new FlagTerm(siFlags, true));
-			}
-			else {
-				logger.debug("This email server does not support RECENT or USER flags. " +
-						"System flag 'Flag.FLAGGED' will be used to prevent duplicates during email fetch.");
-				notFlagged = new NotTerm(new FlagTerm(new Flags(Flag.FLAGGED), true));
-			}
-
-			if (searchTerm == null) {
-				return notFlagged;
-			}
-			else {
-				return new AndTerm(searchTerm, notFlagged);
-			}
-		}
-
 	}
 
 }
