+++ spring-integration-mail/src/main/java/org/springframework/integration/mail/MailReceivingMessageSource.java
@@ -16,21 +16,10 @@
 
 package org.springframework.integration.mail;
 
-import java.util.Arrays;
-import java.util.Queue;
-import java.util.concurrent.ConcurrentLinkedQueue;
-
-import org.jspecify.annotations.Nullable;
-
-import org.springframework.core.log.LogMessage;
-import org.springframework.integration.endpoint.AbstractMessageSource;
-import org.springframework.messaging.MessagingException;
-import org.springframework.util.Assert;
-
 /**
  * {@link org.springframework.integration.core.MessageSource} implementation that
- * delegates to a {@link MailReceiver} to poll a mailbox. Each poll of the mailbox may
- * return more than one message which will then be stored in a queue.
+ * delegates to a {@link org.springframework.integration.mail.inbound.MailReceiver} to poll a mailbox.
+ * Each poll of the mailbox may return more than one message which will then be stored in a queue.
  *
  * @author Jonas Partner
  * @author Mark Fisher
@@ -38,43 +27,15 @@ import org.springframework.util.Assert;
  * @author Oleg Zhurakousky
  * @author Artem Bilan
  * @author Trung Pham
+ *
+ * @deprecated since 7.0 in favor of {@link org.springframework.integration.mail.inbound.MailReceivingMessageSource}
  */
-public class MailReceivingMessageSource extends AbstractMessageSource<Object> {
+@Deprecated(forRemoval = true, since = "7.0")
+public class MailReceivingMessageSource
+		extends org.springframework.integration.mail.inbound.MailReceivingMessageSource {
 
-	private final MailReceiver mailReceiver;
-
-	private final Queue<Object> mailQueue = new ConcurrentLinkedQueue<>();
-
-	public MailReceivingMessageSource(MailReceiver mailReceiver) {
-		Assert.notNull(mailReceiver, "mailReceiver must not be null");
-		this.mailReceiver = mailReceiver;
-	}
-
-	@Override
-	public String getComponentType() {
-		return "mail:inbound-channel-adapter";
-	}
-
-	@Override
-	protected @Nullable Object doReceive() {
-		try {
-			Object mailMessage = this.mailQueue.poll();
-			if (mailMessage == null) {
-				Object[] messages = this.mailReceiver.receive();
-				if (messages != null) {
-					this.mailQueue.addAll(Arrays.asList(messages));
-				}
-				mailMessage = this.mailQueue.poll();
-			}
-			if (mailMessage != null) {
-				this.logger.debug(LogMessage.format("received mail message [%s]", mailMessage));
-				return mailMessage;
-			}
-		}
-		catch (Exception e) {
-			throw new MessagingException("failure occurred while polling for mail", e);
-		}
-		return null;
+	public MailReceivingMessageSource(org.springframework.integration.mail.inbound.MailReceiver mailReceiver) {
+		super(mailReceiver);
 	}
 
 }
