+++ spring-integration-mail/src/main/java/org/springframework/integration/mail/MailSendingMessageHandler.java
@@ -16,27 +16,9 @@
 
 package org.springframework.integration.mail;
 
-import java.util.Arrays;
-import java.util.Objects;
-
-import jakarta.mail.MessagingException;
-import jakarta.mail.internet.MimeMessage;
-import org.jspecify.annotations.Nullable;
-
-import org.springframework.core.io.ByteArrayResource;
-import org.springframework.integration.handler.AbstractMessageHandler;
-import org.springframework.integration.mapping.MessageMappingException;
 import org.springframework.mail.MailMessage;
 import org.springframework.mail.MailSender;
 import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.MimeMailMessage;
-import org.springframework.mail.javamail.MimeMessageHelper;
-import org.springframework.messaging.Message;
-import org.springframework.messaging.MessageHeaders;
-import org.springframework.util.Assert;
-import org.springframework.util.ObjectUtils;
-import org.springframework.util.StringUtils;
 
 /**
  * A {@link org.springframework.messaging.MessageHandler} implementation for sending mail.
@@ -53,11 +35,13 @@ import org.springframework.util.StringUtils;
  * @author Oleg Zhurakousky
  * @author Artem Bilan
  * @author Ma Jiandong
+ *
  * @see MailHeaders
+ *
+ * @deprecated since 7.0 in favor of {@link org.springframework.integration.mail.outbound.MailSendingMessageHandler}
  */
-public class MailSendingMessageHandler extends AbstractMessageHandler {
-
-	private final MailSender mailSender;
+@Deprecated(forRemoval = true, since = "7.0")
+public class MailSendingMessageHandler extends org.springframework.integration.mail.outbound.MailSendingMessageHandler {
 
 	/**
 	 * Create a MailSendingMessageHandler.
@@ -65,152 +49,7 @@ public class MailSendingMessageHandler extends AbstractMessageHandler {
 	 * adapter will delegate.
 	 */
 	public MailSendingMessageHandler(MailSender mailSender) {
-		Assert.notNull(mailSender, "'mailSender' must not be null");
-		this.mailSender = mailSender;
-	}
-
-	@Override
-	public String getComponentType() {
-		return "mail:outbound-channel-adapter";
-	}
-
-	@Override
-	protected final void handleMessageInternal(Message<?> message) {
-		MailMessage mailMessage = convertMessageToMailMessage(message);
-		if (mailMessage instanceof SimpleMailMessage simpleMailMessage) {
-			this.mailSender.send(simpleMailMessage);
-		}
-		else if (mailMessage instanceof MimeMailMessage mimeMailMessage) {
-			Assert.state(this.mailSender instanceof JavaMailSender,
-					"this adapter requires a 'JavaMailSender' to send a 'MimeMailMessage'");
-
-			((JavaMailSender) this.mailSender).send(mimeMailMessage.getMimeMessage());
-		}
-		else {
-			throw new IllegalArgumentException(
-					"Unsupported MailMessage type [" + mailMessage.getClass().getName() + "].");
-		}
-	}
-
-	@SuppressWarnings("unchecked")
-	private MailMessage convertMessageToMailMessage(Message<?> message) {
-		MailMessage mailMessage;
-		Object payload = message.getPayload();
-		if (payload instanceof MimeMessage mimeMessage) {
-			mailMessage = new MimeMailMessage(mimeMessage);
-		}
-		else if (payload instanceof MailMessage mailMsg) {
-			mailMessage = mailMsg;
-		}
-		else if (payload instanceof byte[]) {
-			mailMessage = createMailMessageFromByteArrayMessage((Message<byte[]>) message);
-		}
-		else if (payload instanceof String) {
-			String contentType = (String) message.getHeaders().get(MailHeaders.CONTENT_TYPE);
-			if (StringUtils.hasText(contentType)) {
-				mailMessage = createMailMessageWithContentType((Message<String>) message, contentType);
-			}
-			else {
-				mailMessage = new SimpleMailMessage();
-				mailMessage.setText((String) payload);
-			}
-		}
-		else {
-			throw new IllegalArgumentException("Unable to create MailMessage from payload type ["
-					+ message.getPayload().getClass().getName() + "], " +
-					"expected MimeMessage, MailMessage, byte array or String.");
-		}
-		applyHeadersToMailMessage(mailMessage, message.getHeaders());
-		return mailMessage;
-	}
-
-	private MailMessage createMailMessageWithContentType(Message<String> message, String contentType) {
-		Assert.state(this.mailSender instanceof JavaMailSender,
-				"this adapter requires a 'JavaMailSender' to send a 'MimeMailMessage'");
-
-		MimeMessage mimeMessage = ((JavaMailSender) this.mailSender).createMimeMessage();
-		try {
-			mimeMessage.setContent(message.getPayload(), contentType);
-			return new MimeMailMessage(mimeMessage);
-		}
-		catch (Exception e) {
-			throw new IllegalStateException("Failed to create MimeMessage with contentType: " + contentType, e);
-		}
-	}
-
-	private MailMessage createMailMessageFromByteArrayMessage(Message<byte[]> message) {
-		Assert.state(this.mailSender instanceof JavaMailSender,
-				"this adapter requires a 'JavaMailSender' to send a 'MimeMailMessage'");
-
-		String attachmentFileName = message.getHeaders().get(MailHeaders.ATTACHMENT_FILENAME, String.class);
-		if (attachmentFileName == null) {
-			throw new MessageMappingException(message, "Header '" + MailHeaders.ATTACHMENT_FILENAME
-					+ "' is required when mapping a Message with a byte array payload to a MailMessage.");
-		}
-		Integer multipartMode = message.getHeaders().get(MailHeaders.MULTIPART_MODE, Integer.class);
-		if (multipartMode == null) {
-			multipartMode = MimeMessageHelper.MULTIPART_MODE_MIXED;
-		}
-
-		MimeMessage mimeMessage = ((JavaMailSender) this.mailSender).createMimeMessage();
-		try {
-			MimeMessageHelper helper = new MimeMessageHelper(mimeMessage, multipartMode);
-			helper.addAttachment(attachmentFileName, new ByteArrayResource(message.getPayload()));
-			return new MimeMailMessage(helper);
-		}
-		catch (MessagingException e) {
-			throw new MessageMappingException(message, "failed to create MimeMessage", e);
-		}
-	}
-
-	private void applyHeadersToMailMessage(MailMessage mailMessage, MessageHeaders headers) {
-		String subject = headers.get(MailHeaders.SUBJECT, String.class);
-		if (subject != null) {
-			mailMessage.setSubject(subject);
-		}
-		String[] to = retrieveHeaderValueAsStringArray(headers, MailHeaders.TO);
-		if (to != null) {
-			mailMessage.setTo(to);
-		}
-		if (mailMessage instanceof SimpleMailMessage simpleMailMessage) {
-			Assert.state(!ObjectUtils.isEmpty(simpleMailMessage.getTo()),
-					"No recipient has been provided on the MailMessage or the 'MailHeaders.TO' header.");
-		}
-		String[] cc = retrieveHeaderValueAsStringArray(headers, MailHeaders.CC);
-		if (cc != null) {
-			mailMessage.setCc(cc);
-		}
-		String[] bcc = retrieveHeaderValueAsStringArray(headers, MailHeaders.BCC);
-		if (bcc != null) {
-			mailMessage.setBcc(bcc);
-		}
-		String from = headers.get(MailHeaders.FROM, String.class);
-		if (from != null) {
-			mailMessage.setFrom(from);
-		}
-		String replyTo = headers.get(MailHeaders.REPLY_TO, String.class);
-		if (replyTo != null) {
-			mailMessage.setReplyTo(replyTo);
-		}
-	}
-
-	private String @Nullable [] retrieveHeaderValueAsStringArray(MessageHeaders headers, String key) {
-		Object value = headers.get(key);
-		String[] returnedHeaders = null;
-		if (value != null) {
-			if (value instanceof String[] strArr) {
-				returnedHeaders = Arrays.stream(strArr)
-						.filter(Objects::nonNull)
-						.toArray(String[]::new);
-			}
-			else if (value instanceof String) {
-				returnedHeaders = StringUtils.commaDelimitedListToStringArray((String) value);
-			}
-		}
-		if (returnedHeaders == null || ObjectUtils.isEmpty(returnedHeaders)) {
-			returnedHeaders = null;
-		}
-		return returnedHeaders;
+		super(mailSender);
 	}
 
 }
