+++ spring-integration-mail/src/main/java/org/springframework/integration/mail/config/MailInboundChannelAdapterParser.java
@@ -26,7 +26,7 @@ import org.springframework.beans.factory.xml.ParserContext;
 import org.springframework.integration.config.ExpressionFactoryBean;
 import org.springframework.integration.config.xml.AbstractPollingInboundChannelAdapterParser;
 import org.springframework.integration.config.xml.IntegrationNamespaceUtils;
-import org.springframework.integration.mail.MailReceivingMessageSource;
+import org.springframework.integration.mail.inbound.MailReceivingMessageSource;
 import org.springframework.util.StringUtils;
 import org.springframework.util.xml.DomUtils;
