+++ spring-integration-mail/src/main/java/org/springframework/integration/mail/dsl/MailInboundChannelAdapterSpec.java
@@ -31,8 +31,8 @@ import org.springframework.expression.Expression;
 import org.springframework.integration.dsl.ComponentsRegistration;
 import org.springframework.integration.dsl.MessageSourceSpec;
 import org.springframework.integration.expression.FunctionExpression;
-import org.springframework.integration.mail.AbstractMailReceiver;
-import org.springframework.integration.mail.MailReceivingMessageSource;
+import org.springframework.integration.mail.inbound.AbstractMailReceiver;
+import org.springframework.integration.mail.inbound.MailReceivingMessageSource;
 import org.springframework.integration.mapping.HeaderMapper;
 import org.springframework.integration.support.PropertiesBuilder;
 import org.springframework.util.Assert;
