+++ spring-integration-mail/src/main/java/org/springframework/integration/mail/inbound/MailReceiver.java
@@ -14,22 +14,21 @@
  * limitations under the License.
  */
 
-package org.springframework.integration.mail;
+package org.springframework.integration.mail.inbound;
 
-import jakarta.mail.search.SearchTerm;
+import org.jspecify.annotations.Nullable;
 
 /**
- * Strategy to be used to generate a {@link SearchTerm}.
+ * Strategy interface for receiving mail {@link jakarta.mail.Message Messages}.
  *
- * @author Oleg Zhurakousky
+ * @author Mark Fisher
+ * @author Gary Russell
+ * @author Artem Bilan
  *
- * @since 2.2
- *
- * @see org.springframework.integration.mail.inbound.ImapMailReceiver
- *
- * @deprecated since 7.0 in favor of {@link org.springframework.integration.mail.inbound.SearchTermStrategy}
+ * @since 7.0
  */
-@Deprecated(forRemoval = true, since = "7.0")
-public interface SearchTermStrategy extends org.springframework.integration.mail.inbound.SearchTermStrategy {
+public interface MailReceiver {
+
+	Object @Nullable [] receive() throws jakarta.mail.MessagingException;
 
 }
