+++ spring-integration-mail/src/main/java/org/springframework/integration/mail/inbound/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * The inbound channel adapters support for Mail protocol.
+ */
+@org.jspecify.annotations.NullMarked
+package org.springframework.integration.mail.inbound;
