+++ spring-integration-mail/src/main/java/org/springframework/integration/mail/outbound/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * The outbound channel adapters support for Mail protocol.
+ */
+@org.jspecify.annotations.NullMarked
+package org.springframework.integration.mail.outbound;
