+++ spring-integration-mail/src/test/java/org/springframework/integration/mail/config/InboundChannelAdapterParserTests-pop3Search-context.xml
@@ -14,7 +14,7 @@
 								  search-term-strategy="searchTermStrategy"/>
 
 	<bean id="searchTermStrategy" class="org.mockito.Mockito" factory-method="mock">
-		<constructor-arg value="org.springframework.integration.mail.SearchTermStrategy" type="java.lang.Class"/>
+		<constructor-arg value="org.springframework.integration.mail.inbound.SearchTermStrategy" type="java.lang.Class"/>
 	</bean>
 
 	<!-- COMMON CONFIGURATION -->
