+++ spring-integration-mail/src/test/java/org/springframework/integration/mail/config/InboundChannelAdapterParserTests.java
@@ -28,11 +28,11 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 import org.springframework.integration.endpoint.SourcePollingChannelAdapter;
-import org.springframework.integration.mail.AbstractMailReceiver;
-import org.springframework.integration.mail.ImapIdleChannelAdapter;
-import org.springframework.integration.mail.ImapMailReceiver;
-import org.springframework.integration.mail.Pop3MailReceiver;
-import org.springframework.integration.mail.SearchTermStrategy;
+import org.springframework.integration.mail.inbound.AbstractMailReceiver;
+import org.springframework.integration.mail.inbound.ImapIdleChannelAdapter;
+import org.springframework.integration.mail.inbound.ImapMailReceiver;
+import org.springframework.integration.mail.inbound.Pop3MailReceiver;
+import org.springframework.integration.mail.inbound.SearchTermStrategy;
 import org.springframework.integration.test.util.TestUtils;
 import org.springframework.messaging.MessageChannel;
 import org.springframework.test.annotation.DirtiesContext;
